+++ src/main/java/com/jeff_media/updatechecker/UpdateCheckEvent.java
@@ -1,4 +1,22 @@
-package de.jeff_media.updatechecker;
+/*
+ * Copyright (c) 2022 Alexander Majka (mfnalex), JEFF Media GbR
+ * Website: https://www.jeff-media.com
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.jeff_media.updatechecker;
 
 import org.bukkit.command.CommandSender;
 import org.bukkit.event.Event;
@@ -31,6 +49,7 @@ public class UpdateCheckEvent extends Event {
         }
     }
 
+    @SuppressWarnings("unused")
     public static HandlerList getHandlerList() {
         return HANDLERS;
     }