+++ src/main/java/de/jeff_media/updatechecker/ArtifactVersion.java
@@ -1,5 +1,0 @@
-package de.jeff_media.updatechecker;
-
-interface ArtifactVersion extends Comparable<ArtifactVersion> {
-
-}