+++ src/main/java/de/jeff_media/updatechecker/ThrowingFunction.java
@@ -1,6 +1,0 @@
-package de.jeff_media.updatechecker;
-
-@FunctionalInterface
-interface ThrowingFunction<T,R,E extends Exception> {
-    R apply(T t) throws E;
-}