+++ src/main/java/de/jeff_media/updatechecker/UpdateCheckSuccess.java
@@ -1,9 +1,0 @@
-package de.jeff_media.updatechecker;
-
-/**
- * Represents whether the update check was successful
- */
-public enum UpdateCheckSuccess {
-    SUCCESS,
-    FAIL
-}